fixup! WIP move graph delegate to separate file (after rebase)
